pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "evan-llm-sdk"
version = "0.1.0"
description = "A Python package for generating text using large language models."
readme = "README.md"
requires-python = ">=3.10"
license = { file = "LICENSE" }
authors = [{ name="Evan", email="e2002florespulido@gmail.com" }]
keywords = ["logging", "colorlog", "structured-logging", "trace"]
dependencies = [
  "annotated-types>=0.7.0",
  "anyio>=4.12.1",
  "certifi>=2026.1.4",
  "cffi>=2.0.0",
  "charset-normalizer>=3.4.4",
  "colorlog>=6.10.1",
  "cryptography>=46.0.4",
  "distro>=1.9.0",
  "dotenv>=0.9.9",
  "exceptiongroup>=1.3.1",
  "google-ai-generativelanguage>=0.6.15",
  "google-api-core>=2.29.0",
  "google-api-python-client>=2.189.0",
  "google-auth>=2.48.0",
  "google-auth-httplib2>=0.3.0",
  "google-cloud-core>=2.5.0",
  "google-cloud-storage>=3.9.0",
  "google-crc32c>=1.8.0",
  "google-genai>=1.62.0",
  "google-generativeai>=0.8.6",
  "google-resumable-media>=2.8.0",
  "googleapis-common-protos>=1.72.0",
  "grpcio>=1.76.0",
  "grpcio-status>=1.71.2",
  "h11>=0.16.0",
  "httpcore>=1.0.9",
  "httplib2>=0.31.2",
  "httpx>=0.28.1",
  "idna>=3.11",
  "iniconfig>=2.3.0",
  "packaging>=26.0",
  "pluggy>=1.6.0",
  "proto-plus>=1.27.1",
  "protobuf>=5.29.6",
  "pyasn1>=0.6.2",
  "pyasn1_modules>=0.4.2",
  "pycparser>=3.0",
  "pydantic>=2.12.5",
  "pydantic-settings>=2.12.0",
  "pydantic_core>=2.41.5",
  "Pygments>=2.19.2",
  "pyparsing>=3.3.2",
  "pytest>=9.0.2",
  "python-dotenv>=1.2.1",
  "requests>=2.32.5",
  "rsa>=4.9.1",
  "sniffio>=1.3.1",
  "tenacity>=9.1.3",
  "tomli>=2.4.0",
  "tqdm>=4.67.3",
  "typing-inspection>=0.4.2",
  "typing_extensions>=4.15.0",
  "uritemplate>=4.2.0",
  "urllib3>=2.6.3",
  "websockets>=15.0.1",
]

[project.urls]
Homepage = "https://github.com/EvanFloresLv/LLM-Generate.git"
Repository = "https://github.com/EvanFloresLv/LLM-Generate.git"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-cov>=5.0",
  "pytest-asyncio>=0.23",
  "mypy>=1.8",
  "ruff>=0.3",
  "respx>=0.20",
]

[project.entry-points."llm_sdk.providers"]
gemini = "llm_sdk.providers.gemini.plugin:GeminiProviderFactory"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q --disable-warnings --maxfail=1"
asyncio_mode = "auto"

[tool.mypy]
python_version = "3.11"
strict = true
mypy_path = ["src"]
warn_return_any = true
warn_unused_ignores = true
no_implicit_optional = true

[tool.ruff]
line-length = 100
target-version = "py311"